+++ src/components/WidgetComponents/SimpleSearch/SimpleSearch.js
@@ -7,7 +7,8 @@ import { FormattedMessage } from 'react-intl';
 import moment from 'moment';
 
 import { useQuery } from 'react-query';
-import { useOkapiKy } from '@folio/stripes/core';
+import { useOkapiKy, useStripes } from '@folio/stripes/core';
+
 import { Badge } from '@folio/stripes/components';
 
 import pathBuilder from './simpleSearchPathBuilder';
@@ -33,9 +34,11 @@ const SimpleSearch = ({
   const columns = columnParser({ widgetDef, widgetConf });
 
   const ky = useOkapiKy();
+  // We need to pass the stripes object into the pathBuilder, so it can use that for currentUser token
+  const stripes = useStripes();
   const { data, dataUpdatedAt, refetch } = useQuery(
     ['ui-dashboard', 'simpleSearch', widget.id],
-    () => ky(pathBuilder(widgetDef, widgetConf)).json()
+    () => ky(pathBuilder(widgetDef, widgetConf, stripes)).json()
   );
 
   const timestamp = dataUpdatedAt ? moment(dataUpdatedAt).format('hh:mm a') : '';