+++ src/components/WidgetComponents/SimpleSearch/simpleSearchPathBuilder.js
@@ -1,4 +1,6 @@
-const simpleSearchPathBuilder = (widgetDef, widgetConf) => {
+import tokens from '../../../tokens';
+
+const simpleSearchPathBuilder = (widgetDef, widgetConf, stripes) => {
   const {
     baseUrl,
     filters: {
@@ -83,11 +85,11 @@ const simpleSearchPathBuilder = (widgetDef, widgetConf) => {
       // Then take each of the rules within the filter, and OR them together with the correct comparators
       const { rules } = f;
       rules.forEach((r, ind) => {
-        if (r.isNull) {
+        if (r.comparator === 'isNull' || r.comparator === 'isNotNull') {
           // If we're allowing null the filterString is slightly different
-          specificFilterString += `${filterPath} isNull`;
+          specificFilterString += `${filterPath} ${r.comparator}`;
         } else {
-          specificFilterString += `${filterPath}${r.comparator}${r.filterValue}`;
+          specificFilterString += `${filterPath}${r.comparator}${tokens(r.filterValue, stripes)}`;
         }
 
         if (ind !== rules.length - 1) {