+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change history for ui-dashboard
 
 ## 1.0.0 In progress
+* Added the ability to tokenise currentDate and currentUser
+* ERM-1562 Dynamic form rendering for simpleSearch widgets
 * ERM-1561 App displays SimpleSearch results in formatted table
 * New app created with stripes-cli
 * Updated to use Babel 7.